LPS-55833 Move edit configuration action to portlet
